+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.2.21
 
-  * Fix updating badge data to a blank string incorrectly returns false when the update was made into OSM
+  * Fix updating requirement data of a Osm::Badge::Data to a blank string incorrectly returns false when the update was made into OSM
+  * Fix updating additional details for a Osm::Member to a blank string incorrectly returns false when the update was made into OSM
 
 ## Version 1.2.20